requirements.txt
pimoroni-sgp30
smbus2
RPi.GPIO
pms5003
pimoroni-bme280
influxdb
